Monitor timeout in build
We have been bitten by execution timeout for a long time. Now we set a hourly timer when
integration tests start, and print all java processes' stack traces on the machine.
